+++ src/main/java/roomescape/common/exception/RoomEscapeException.java
@@ -1,7 +1,6 @@
 package roomescape.common.exception;
 
 import lombok.Getter;
-import org.springframework.http.HttpStatus;
 
 @Getter
 public class RoomEscapeException extends RuntimeException{
