+++ src/main/java/roomescape/date/controller/ReservationDateController.java
@@ -6,7 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.date.dto.response.ReservationDateDetailDto;
+import roomescape.date.controller.dto.response.ReservationDateDetailDto;
 import roomescape.date.service.ReservationDateService;
 
 @RestController
