+++ src/main/java/roomescape/date/controller/dto/request/ReservationDateStatusUpdateDto.java
@@ -1,8 +1,8 @@
-package roomescape.theme.dto.request;
+package roomescape.date.controller.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 
-public record ThemeActiveUpdateDto(
+public record ReservationDateStatusUpdateDto(
         @NotNull(message = "isActive는 필수 입력입니다.")
         Boolean isActive
 ) {
