+++ src/main/java/roomescape/reservation/controller/dto/response/ReservationDetailDto.java
@@ -1,11 +1,11 @@
-package roomescape.reservation.dto.response;
+package roomescape.reservation.controller.dto.response;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationStatus;
-import roomescape.theme.dto.response.ThemeDetailDto;
+import roomescape.theme.controller.dto.response.ThemeDetailDto;
 
 public record ReservationDetailDto(
         Long id,
