+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.theme.dto.response.ThemeDetailDto;
+import roomescape.theme.controller.dto.response.ThemeDetailDto;
 import roomescape.theme.service.ThemeService;
 
 @RestController
