+++ src/main/java/roomescape/theme/controller/dto/request/ThemeActiveUpdateDto.java
@@ -1,8 +1,8 @@
-package roomescape.date.dto.request;
+package roomescape.theme.controller.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 
-public record ReservationDateStatusUpdateDto(
+public record ThemeActiveUpdateDto(
         @NotNull(message = "isActive는 필수 입력입니다.")
         Boolean isActive
 ) {
