+++ src/main/java/roomescape/time/controller/ReservationTimeController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.time.dto.response.ReservationTimeDetailDto;
+import roomescape.time.controller.dto.response.ReservationTimeDetailDto;
 import roomescape.time.service.ReservationTimeService;
 
 @RestController
