+++ src/main/java/roomescape/time/controller/dto/request/ReservationTimeStatusUpdateDto.java
@@ -1,4 +1,4 @@
-package roomescape.time.dto.request;
+package roomescape.time.controller.dto.request;
 
 import jakarta.validation.constraints.NotNull;
